✨ Add Project resolver and service
