+++ server/src/project/project.resolver.ts
@@ -1,13 +1,29 @@
 import { Args, Query, Resolver } from '@nestjs/graphql';
 import {
+  GetAllProjectsResponseSchema,
+  GetCategoryNameAndIdResponeSchema,
   GetProjectByCategoryResponseSchema,
   GetProjectByCategorySchema,
 } from 'src/schema/service/project.service-schema';
-import { ProjectService } from 'src/service/project.service';
+import { ProjectService } from './project.service';
 
 @Resolver()
 export class ProjectResolver {
   constructor(private projectService: ProjectService) {}
+
+  @Query(() => GetAllProjectsResponseSchema)
+  async getAllProjects(): Promise<GetAllProjectsResponseSchema> {
+    const projects = await this.projectService.getAllProjects();
+    return {
+      project: projects,
+      listLength: projects.length,
+    };
+  }
+
+  @Query(() => [GetCategoryNameAndIdResponeSchema])
+  async getCategoryNameAndId(): Promise<GetCategoryNameAndIdResponeSchema[]> {
+    return this.projectService.getCategoryNameAndId(); // 서비스에서 호출
+  }
 
   @Query(() => GetProjectByCategoryResponseSchema)
   async getProjectByCategory(
@@ -23,14 +39,3 @@ export class ProjectResolver {
     };
   }
 }
-
-//   @Mutation(() => UpdateBookResponseSchema)
-//   async updateBook(@Args() args: UpdateBookRequestSchema) {
-//     const { book } = args;
-
-//     const updatedBook = await this.bookService.updateBook(book);
-
-//     return {
-//       book: updatedBook,
-//     };
-//   }
