--- server/src/resolver/index.ts
@@ -1,3 +1,0 @@
-import { ProjectResolver } from './project.resolver';
-
-export const rootResolvers = [ProjectResolver];
